refactor python client structure
