+++ client-py/src/SessionExample.py
@@ -16,20 +16,10 @@
 # under the License.
 #
 
-import sys
 # Uncomment the following line to use apache-iotdb module installed by pip3
 
-# for debug
-# sys.path.append("./utils")
-# from IoTDBConstants import *
-# from Tablet import Tablet
-# from Session import Session
-
-# for package
-from iotdb_session.utils.IoTDBConstants import *
-from iotdb_session.utils.Tablet import Tablet
-from iotdb_session.Session import Session
-
+from iotdb.Session import Session
+from iotdb.utils import Tablet
 
 # creating session connection.
 ip = "127.0.0.1"
